# src/main/java/org/mars_group/gisimport/controller/FileController.java
package org.mars_group.gisimport.controller;

import it.geosolutions.geoserver.rest.GeoServerRESTPublisher;
import org.apache.commons.io.FileUtils;
import org.mars_group.core.ImportState;
import org.mars_group.gisimport.exceptions.GisImportException;
import org.mars_group.gisimport.util.GeoServer;
import org.mars_group.metadataclient.MetadataClient;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpMethod;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.util.FileCopyUtils;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.client.RestTemplate;

import java.io.*;
import java.net.MalformedURLException;

import static org.junit.Assert.assertTrue;

@RestController
class FileController {

    private static final String uploadDir = "upload-dir";

    private final RestTemplate restTemplate;
    private final GeoServerImport geoServerImport;
    private final GeoServer geoServer;
    private final GeoServerExport geoServerExport;

    @Autowired
    public FileController(RestTemplate restTemplate, GeoServerImport geoServerImport, GeoServer geoServer, GeoServerExport geoServerExport) {
        this.restTemplate = restTemplate;
        this.geoServerImport = geoServerImport;
        this.geoServer = geoServer;
        this.geoServerExport = geoServerExport;
    }

    /**
     * import Geo files
     *
     * @param dataId   generated by file-controller
     * @param title    specified by user
     * @param filename the name of the file, the user uploaded
     * @return status message
     */
    @ResponseBody
    @RequestMapping(method = RequestMethod.POST, value = "/gis")
    public ResponseEntity<String> postFile(@RequestParam String dataId, @RequestParam String title, @RequestParam String filename) {
        String uri = "http://file-svc/files/";

        return restTemplate.execute(uri + dataId, HttpMethod.GET, null, response -> {
            try {
                String specificUploadDir = uploadDir + File.separator + dataId;

                saveFileToDisk(response.getBody(), specificUploadDir, filename);
                return startImport(title, filename, dataId, specificUploadDir);

            } catch (GisImportException e) {
                e.printStackTrace();
                return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
            }
        });
    }

    /**
     * Download files
     *
     * @param dataId id created during import
     * @return relative uri to the file
     */
    @ResponseBody
    @RequestMapping(method = RequestMethod.GET, value = "/gis/{dataId}")
    public ResponseEntity<String> getFile(@PathVariable("dataId") String dataId) {
        return new ResponseEntity<>(geoServerExport.getUriFromDataId(dataId).toString(), HttpStatus.OK);
    }

    @ResponseBody
    @RequestMapping(method = RequestMethod.DELETE, value = "/gis/{dataId}")
    public ResponseEntity<String> deleteFile(@PathVariable("dataId") String dataId) throws MalformedURLException, GisImportException {
        GeoServerRESTPublisher publisher = geoServer.getPublisher();
        boolean result = publisher.removeWorkspace(dataId, true);

        if (result) {
            return new ResponseEntity<>("deleted", HttpStatus.OK);
        } else {
            return new ResponseEntity<>("Error inside the geoserver", HttpStatus.INTERNAL_SERVER_ERROR);

        }
    }

    private ResponseEntity<String> startImport(String title, String filename, String dataId, String specificUploadDir) {
        MetadataClient metadataClient = new MetadataClient(restTemplate);
        metadataClient.setState(dataId, ImportState.PROCESSING);

        try {
            // START THE IMPORT
            geoServerImport.handleImport(specificUploadDir, filename, dataId, title);
        } catch (GisImportException | MalformedURLException e) {
            e.printStackTrace();
            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
        }

        metadataClient.setState(dataId, ImportState.FINISHED);

        try {
            FileUtils.deleteDirectory(new File(specificUploadDir));
        } catch (IOException e) {
            e.printStackTrace();
        }

        return new ResponseEntity<>(dataId, HttpStatus.OK);
    }

    private void saveFileToDisk(InputStream file, String specificUploadDir, String filename) throws GisImportException {
        try {
            if (!new File(uploadDir).exists()) {
                assertTrue(new File(uploadDir).mkdir());
            }

            if (!new File(specificUploadDir).exists()) {
                assertTrue(new File(specificUploadDir).mkdir());
            }

            // save file
            File f = new File(specificUploadDir + File.separator + filename);
            BufferedOutputStream stream = new BufferedOutputStream(new FileOutputStream(f));
            FileCopyUtils.copy(file, stream);
            stream.close();
        } catch (Exception e) {
            e.printStackTrace();
            throw new GisImportException(e.getMessage());
        }
    }

}